레이아웃 파일에서 SplashScreen 숨기기 및 RevenueCat 설정을 위한 커스텀 훅 추가, 코드 구조 개선
