+++ app/_layout.tsx
@@ -4,14 +4,13 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { Stack } from 'expo-router';
 import * as SplashScreen from 'expo-splash-screen';
 import { OverlayProvider } from 'overlay-kit';
-import { useEffect } from 'react';
-import { Platform } from 'react-native';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
-import Purchases, { LOG_LEVEL } from 'react-native-purchases';
+import { KeyboardProvider } from 'react-native-keyboard-controller';
 import { DevToolsBubble } from 'react-native-react-query-devtools';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { StyleSheet } from 'react-native-unistyles';
 import { AuthLoaded } from '~/shared/auth';
+import { useHideSplashScreen, useRevenueCat } from '~/shared/lib';
 import { Toast } from '~/shared/ui/toast';
 
 const queryClient = new QueryClient({
@@ -35,34 +34,9 @@ SplashScreen.preventAutoHideAsync();
  * - https://github.com/expo/expo/issues/33673
  */
 export default function Layout() {
-  useEffect(() => {
-    const hideSplash = async () => {
-      await SplashScreen.hideAsync();
-    };
+  useHideSplashScreen();
 
-    hideSplash();
-  }, []);
-
-  useEffect(() => {
-    Purchases.setLogLevel(LOG_LEVEL.VERBOSE);
-
-    if (
-      !process.env.EXPO_PUBLIC_REVENUECAT_PROJECT_APPLE_API_KEY ||
-      !process.env.EXPO_PUBLIC_REVENUECAT_PROJECT_GOOGLE_API_KEY
-    ) {
-      throw new Error('RevenueCat API key is not set');
-    }
-
-    if (Platform.OS === 'ios') {
-      Purchases.configure({
-        apiKey: process.env.EXPO_PUBLIC_REVENUECAT_PROJECT_APPLE_API_KEY,
-      });
-    } else if (Platform.OS === 'android') {
-      Purchases.configure({
-        apiKey: process.env.EXPO_PUBLIC_REVENUECAT_PROJECT_GOOGLE_API_KEY,
-      });
-    }
-  }, []);
+  useRevenueCat();
 
   return (
     <GestureHandlerRootView style={styles.container}>
@@ -70,19 +44,21 @@ export default function Layout() {
         <AuthLoaded>
           <OverlayProvider>
             <SafeAreaView edges={['left', 'right']} style={styles.container}>
-              <BottomSheetModalProvider>
-                <Stack
-                  screenOptions={{
-                    headerShown: false,
-                    contentStyle: styles.container,
-                  }}>
-                  <Stack.Screen name="index" options={{ title: 'Home' }} />
-                  <Stack.Screen name="(authorized)" />
-                  <Stack.Screen name="(unauthorized)" />
-                </Stack>
-                <Toast />
-                {__DEV__ && <DevToolsBubble queryClient={queryClient} />}
-              </BottomSheetModalProvider>
+              <KeyboardProvider>
+                <BottomSheetModalProvider>
+                  <Stack
+                    screenOptions={{
+                      headerShown: false,
+                      contentStyle: styles.container,
+                    }}>
+                    <Stack.Screen name="index" options={{ title: 'Home' }} />
+                    <Stack.Screen name="(authorized)" />
+                    <Stack.Screen name="(unauthorized)" />
+                  </Stack>
+                  <Toast />
+                  {__DEV__ && <DevToolsBubble queryClient={queryClient} />}
+                </BottomSheetModalProvider>
+              </KeyboardProvider>
             </SafeAreaView>
           </OverlayProvider>
         </AuthLoaded>
