+++ shared/lib/index.ts
@@ -1,4 +1,6 @@
 export * from './use-active-route';
 export { default as hexToRgba } from './hex-to-rgba';
-export { default as useKeyboard } from './use-keyboard';
 export { default as formatZonedTime } from './format-zoned-time';
+export { default as useKeyboard } from './use-keyboard';
+export { default as useRevenueCat } from './use-revenuecat';
+export { default as useHideSplashScreen } from './use-hide-splash-screen';
